Processing MAAP files
In this Jupyter Notebook, we process the MAAP data for Zeppelin Observatory:

The data comes in two different data sets i.e. 2014 - 2018 and 2018 - 2023.

For 2014 - 2018:

In order to clean we put a minimum flow rate of 900litres per hour.

For 2014 - 2018 we analysed the flags present which were as follows: '000010' '000019' '000012' '000000' '00001A' 0 2 200 89 82 10. We accepted just the flags '000000', 0

Its is assumed that STP is applied to the MAAP data between 2014 and 2018 because the only columns available are as follows: 'date', 'time', 'status', 'concentration of black carbon', 'mass of black carbon', 'air flow rate', 'last value', 'mean values of the concentration\n of black carbon over 1h', '3h','24h'. Hence no, temperatute or pressure of the sample.

We had a look at the distribution of the data (i.e. histogram)

For 2014 to 2018 data we had to convert the mass concentration (given in the files) to absoprtion coefficients. This was done use a MAC of 6.6

Detection limits no applied

Hourly means calculated

For 2018 - 2023:

The maximum values i.e. default flag values are replaced with np.nan and then removed.
for 2018 - 2023 STP is already applied this is given by the flags with '0x0200'. 0x0200 signifies that STP correction is applied.
Status Flags present: 'FFFFFFFFFF' '200' 200 '0000000200' '0000100200' '0000300200'
Flags not allowed 'FFFFFFFFFF', '0000100200'
Flow limit of 900 litres per hour applied
Detection limits no applied
Hourly means calculated
All data:

The data sets are concatenated and saved
the saved file is named based on whether to keep all the data points i.e. so-called extreme values (below detection limit)
the file we want is: 'C:\Users\DominicHeslinRees\Documents\Analysis\absorption\MAAP\hourly_MAAP_mean_extremes.dat'